Rename plagiarism classes to better suit its purpose
