+++ rest/src/main/kotlin/org/flaxo/rest/api/PlagiarismController.kt
@@ -4,7 +4,7 @@ import org.apache.commons.collections4.map.PassiveExpiringMap
 import org.flaxo.common.data.PlagiarismReport
 import org.flaxo.common.data.plagiarism.PlagiarismGraph
 import org.flaxo.common.of
-import org.flaxo.rest.manager.plagiarism.PlagiarismManager
+import org.flaxo.rest.manager.plag.PlagiarismManager
 import org.flaxo.rest.manager.response.Response
 import org.flaxo.rest.manager.response.ResponseManager
 import org.springframework.security.access.prepost.PreAuthorize
